+++ src/algorithms/heuristics/heuristics.cpp
@@ -137,6 +137,99 @@ inline void seed_route(const Input& input,
   }
 }
 
+template <class Route> struct UnassignedCosts {
+  const Vehicle& vehicle;
+  Cost max_edge_cost;
+  std::vector<Cost> min_route_to_unassigned;
+  std::vector<Cost> min_unassigned_to_route;
+
+  UnassignedCosts(const Input& input,
+                  Route& route,
+                  const std::set<Index>& unassigned)
+    : vehicle(input.vehicles[route.vehicle_rank]),
+      max_edge_cost(utils::max_edge_eval(input, vehicle, route.route).cost),
+      min_route_to_unassigned(input.jobs.size(),
+                              std::numeric_limits<Cost>::max()),
+      min_unassigned_to_route(input.jobs.size(),
+                              std::numeric_limits<Cost>::max()) {
+    for (const auto job_rank : unassigned) {
+      const auto unassigned_job_index = input.jobs[job_rank].index();
+
+      if (vehicle.has_start()) {
+        const auto start_to_job =
+          vehicle.eval(vehicle.start.value().index(), unassigned_job_index)
+            .cost;
+        min_route_to_unassigned[job_rank] = start_to_job;
+      }
+
+      if (vehicle.has_end()) {
+        const auto job_to_end =
+          vehicle.eval(unassigned_job_index, vehicle.end.value().index()).cost;
+        min_unassigned_to_route[job_rank] = job_to_end;
+      }
+
+      for (const auto j : route.route) {
+        const auto job_index = input.jobs[j].index();
+
+        const auto job_to_unassigned =
+          vehicle.eval(job_index, unassigned_job_index).cost;
+        min_route_to_unassigned[job_rank] =
+          std::min(min_route_to_unassigned[job_rank], job_to_unassigned);
+
+        const auto unassigned_to_job =
+          vehicle.eval(unassigned_job_index, job_index).cost;
+        min_unassigned_to_route[job_rank] =
+          std::min(min_unassigned_to_route[job_rank], unassigned_to_job);
+      }
+    }
+  }
+
+  double get_insertion_lower_bound(Index j) {
+    return static_cast<double>(min_route_to_unassigned[j] +
+                               min_unassigned_to_route[j] - max_edge_cost);
+  }
+
+  double get_pd_insertion_lower_bound(const Input& input, Index p) {
+    assert(input.jobs[p].type == JOB_TYPE::PICKUP);
+
+    // Situation where pickup and delivery are not inserted in a row.
+    const auto apart_insertion = static_cast<double>(
+      min_route_to_unassigned[p] + min_unassigned_to_route[p] +
+      min_route_to_unassigned[p + 1] + min_unassigned_to_route[p + 1] -
+      2 * max_edge_cost);
+
+    // Situation where delivery is inserted next to the pickup.
+    const auto next_insertion = static_cast<double>(
+      min_route_to_unassigned[p] + min_unassigned_to_route[p + 1] +
+      vehicle.eval(input.jobs[p].index(), input.jobs[p + 1].index()).cost -
+      max_edge_cost);
+
+    return std::min(apart_insertion, next_insertion);
+  }
+
+  void update_max_edge(const Input& input, Route& route) {
+    max_edge_cost = utils::max_edge_eval(input, vehicle, route.route).cost;
+  }
+
+  void update_min_costs(const Input& input,
+                        const std::set<Index>& unassigned,
+                        Index inserted_index) {
+    for (const auto j : unassigned) {
+      const auto unassigned_job_index = input.jobs[j].index();
+
+      const auto to_unassigned =
+        vehicle.eval(inserted_index, unassigned_job_index).cost;
+      min_route_to_unassigned[j] =
+        std::min(min_route_to_unassigned[j], to_unassigned);
+
+      const auto from_unassigned =
+        vehicle.eval(unassigned_job_index, inserted_index).cost;
+      min_unassigned_to_route[j] =
+        std::min(min_unassigned_to_route[j], from_unassigned);
+    }
+  }
+};
+
 template <class Route>
 inline Eval fill_route(const Input& input,
                        Route& route,
@@ -148,6 +241,9 @@ inline Eval fill_route(const Input& input,
 
   const bool init_route_is_empty = route.empty();
   Eval route_eval = utils::route_eval_for_vehicle(input, v_rank, route.route);
+
+  // Store bounds to be able to cut out some loops.
+  UnassignedCosts unassigned_costs(input, route, unassigned);
 
   bool keep_going = true;
   while (keep_going) {
@@ -173,6 +269,14 @@ inline Eval fill_route(const Input& input,
 
       if (current_job.type == JOB_TYPE::SINGLE &&
           route.size() + 1 <= vehicle.max_tasks) {
+
+        if (best_cost < unassigned_costs.get_insertion_lower_bound(job_rank) -
+                          lambda * static_cast<double>(regrets[job_rank])) {
+          // Bypass going through whole route if we're sure insertion
+          // cost is not good enough.
+          continue;
+        }
+
         for (Index r = 0; r <= route.size(); ++r) {
           const auto current_eval =
             utils::addition_cost(input, job_rank, vehicle, route.route, r);
@@ -198,6 +302,15 @@ inline Eval fill_route(const Input& input,
 
       if (current_job.type == JOB_TYPE::PICKUP &&
           route.size() + 2 <= vehicle.max_tasks) {
+
+        if (best_cost <
+            unassigned_costs.get_pd_insertion_lower_bound(input, job_rank) -
+              lambda * static_cast<double>(regrets[job_rank])) {
+          // Bypass going through whole route if we're sure insertion
+          // cost is not good enough.
+          continue;
+        }
+
         // Pre-compute cost of addition for matching delivery.
         std::vector<Eval> d_adds(route.route.size() + 1);
         std::vector<unsigned char> valid_delivery_insertions(
@@ -310,12 +423,16 @@ inline Eval fill_route(const Input& input,
     }
 
     if (best_cost < std::numeric_limits<double>::max()) {
-      if (input.jobs[best_job_rank].type == JOB_TYPE::SINGLE) {
+      const auto& best_job = input.jobs[best_job_rank];
+      if (best_job.type == JOB_TYPE::SINGLE) {
         route.add(input, best_job_rank, best_r);
         unassigned.erase(best_job_rank);
         keep_going = true;
+
+        unassigned_costs.update_max_edge(input, route);
+        unassigned_costs.update_min_costs(input, unassigned, best_job.index());
       }
-      if (input.jobs[best_job_rank].type == JOB_TYPE::PICKUP) {
+      if (best_job.type == JOB_TYPE::PICKUP) {
         std::vector<Index> modified_with_pd;
         modified_with_pd.reserve(best_delivery_r - best_pickup_r + 2);
         modified_with_pd.push_back(best_job_rank);
@@ -334,6 +451,13 @@ inline Eval fill_route(const Input& input,
         unassigned.erase(best_job_rank);
         unassigned.erase(best_job_rank + 1);
         keep_going = true;
+
+        unassigned_costs.update_max_edge(input, route);
+        unassigned_costs.update_min_costs(input, unassigned, best_job.index());
+        unassigned_costs
+          .update_min_costs(input,
+                            unassigned,
+                            input.jobs[best_job_rank + 1].index());
       }
 
       route_eval += best_eval;